refactored mobile menu root as separate component
